requirements.txt
Django~=3.2.17
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
pillow==10.2.0
celery[redis]
psycopg2
pandas

python-dotenv~=0.21.1